# docs/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Zidnee API Documentation",
    "version": "1.0.0",
    "description": "CRM + Student Lifecycle + Follow-up Automation platform API",
    "contact": {
      "name": "Zidnee Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.zidnee.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/leads": {
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Create a new lead",
        "description": "Create a new lead (requires LEAD_CREATE permission)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "level": {
                    "type": "string"
                  },
                  "demoRequired": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "phone",
                  "level"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lead created successfully"
          }
        }
      },
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List all leads",
        "description": "Retrieve a paginated list of leads (requires LEAD_READ permission)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of leads"
          }
        }
      }
    },
    "/api/leads/for-demo": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List pending demo leads",
        "description": "Get leads pending demo assignment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of leads pending demo"
          }
        }
      }
    },
    "/api/leads/demo-requests": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List demo requests",
        "description": "Get all demo requests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of demo requests"
          }
        }
      }
    },
    "/api/leads/admissions": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "List admission leads",
        "description": "Get leads in admission stage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of admission leads"
          }
        }
      }
    },
    "/api/leads/{leadId}": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Get lead by ID",
        "description": "Retrieve a specific lead by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Lead details"
          }
        }
      },
      "patch": {
        "tags": [
          "Leads"
        ],
        "summary": "Update lead",
        "description": "Update lead information",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lead updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Leads"
        ],
        "summary": "Delete lead",
        "description": "Delete a lead (requires LEAD_DELETE permission)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Lead deleted successfully"
          }
        }
      }
    },
    "/api/leads/{leadId}/follow-up/postpone": {
      "patch": {
        "tags": [
          "Leads"
        ],
        "summary": "Postpone follow-up",
        "description": "Postpone the next follow-up date for a lead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nextFollowUpAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Follow-up postponed successfully"
          }
        }
      }
    },
    "/api/leads/{leadId}/demo/request": {
      "patch": {
        "tags": [
          "Leads"
        ],
        "summary": "Request demo for lead",
        "description": "Request a demo session for a lead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Demo requested successfully"
          }
        }
      }
    },
    "/api/leads/{leadId}/form-link": {
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Generate form link",
        "description": "Generate a form submission link for a lead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Form link generated"
          }
        }
      }
    },
    "/api/leads/{leadId}/activities": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Get lead activities",
        "description": "Get activity log for a lead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Lead activities"
          }
        }
      }
    },
    "/form/{leadId}/submit": {
      "post": {
        "tags": [
          "Lead Forms"
        ],
        "summary": "Submit lead form",
        "description": "Submit lead form (public endpoint)",
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Form submitted successfully"
          }
        }
      }
    },
    "/form/{leadId}/validate": {
      "get": {
        "tags": [
          "Lead Forms"
        ],
        "summary": "Validate form link",
        "description": "Validate if a form link is valid (public endpoint)",
        "parameters": [
          {
            "in": "path",
            "name": "leadId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Form link is valid"
          }
        }
      }
    },
...
  },
...
}